+++ index.ts
@@ -124,7 +124,7 @@ export type {SchemaType} from './schema';
 export type {SliderSchema} from './schema';
 export type {StringDateSchema} from './schema';
 export type {StringDateTimeSchema} from './schema';
-export type {StringSchema} from './schema';
+export type {BaseStringSchema} from './schema';
 export type {StringTimeSchema} from './schema';
 export {ValueHintType} from './schema';
 export {ValueType} from './schema';