+++ schema.ts
@@ -172,19 +172,19 @@ export interface ScaleSchema extends BaseNumberSchema {
   icon?: ScaleIconSet;
 }
 
-export interface StringDateSchema extends StringSchema {
+export interface StringDateSchema extends BaseStringSchema {
   codaType: ValueHintType.Date;
   // A Moment date format string, such as 'MMM D, YYYY', that corresponds to a supported Coda date column format.
   format?: string;
 }
 
-export interface StringTimeSchema extends StringSchema {
+export interface StringTimeSchema extends BaseStringSchema {
   codaType: ValueHintType.Time;
   // A Moment time format string, such as 'HH:mm:ss', that corresponds to a supported Coda time column format.
   format?: string;
 }
 
-export interface StringDateTimeSchema extends StringSchema {
+export interface StringDateTimeSchema extends BaseStringSchema {
   codaType: ValueHintType.DateTime;
   // A Moment date format string, such as 'MMM D, YYYY', that corresponds to a supported Coda date column format.
   dateFormat?: string;
@@ -199,15 +199,21 @@ export enum DurationUnit {
   Seconds = 'seconds',
 }
 
-export interface DurationSchema extends StringSchema<ValueHintType.Duration> {
+export interface DurationSchema extends BaseStringSchema<ValueHintType.Duration> {
   precision?: number;
   maxUnit?: DurationUnit;
 }
 
-export interface StringSchema<T extends StringHintTypes = StringHintTypes> extends BaseSchema {
+export interface BaseStringSchema<T extends StringHintTypes = StringHintTypes> extends BaseSchema {
   type: ValueType.String;
   codaType?: T;
 }
+
+export type StringSchema = BaseStringSchema | 
+                           StringDateSchema | 
+                           StringTimeSchema | 
+                           StringDateTimeSchema | 
+                           DurationSchema;
 
 export interface ArraySchema<T extends Schema = Schema> extends BaseSchema {
   type: ValueType.Array;